Implementation of the eccentricity-based lifting of graphs to simplicial complexes
