+++ modules/transforms/data_transform.py
@@ -15,8 +15,8 @@ from modules.transforms.liftings.graph2hypergraph.knn_lifting import (
 from modules.transforms.liftings.graph2simplicial.clique_lifting import (
     SimplicialCliqueLifting,
 )
-from modules.transforms.liftings.graph2simplicial.graph_induced_lifting import (
-    SimplicialGraphInducedLifting,
+from modules.transforms.liftings.graph2simplicial.eccentricity_lifting import (
+    SimplicialEccentricityLifting,
 )
 
 TRANSFORMS = {
@@ -24,7 +24,7 @@ TRANSFORMS = {
     "HypergraphKNNLifting": HypergraphKNNLifting,
     # Graph -> Simplicial Complex
     "SimplicialCliqueLifting": SimplicialCliqueLifting,
-    "SimplicialGraphInducedLifting": SimplicialGraphInducedLifting,
+    "SimplicialEccentricityLifting": SimplicialEccentricityLifting,
     # Graph -> Cell Complex
     "CellCycleLifting": CellCycleLifting,
     # Feature Liftings
